sync version 5.0.1

+++ files/common/usr/lib/ddmesh/ddmesh-firmware-autoupdate.sh
@@ -2,18 +2,20 @@
 # checks or runs an firmware update. downloads firmware
 # arguments: <run | check>
 
-. /lib/functions.sh
-. /lib/upgrade/common.sh
-. /lib/upgrade/platform.sh
-eval $(cat /etc/openwrt_release)
-
-FIRMWARE="$(/usr/lib/ddmesh/ddmesh-get-firmware-name.sh)"
-URL_RELEASE="$(uci get credentials.url.firmware_download_release)"
-URL_ARCH="$DISTRIB_TARGET"
-FIRMWARE_FILE="/tmp/firmware.bin"
 TAG="AutoFirmwareUpdate"
+FIRMWARE_FILE="/tmp/firmware.bin"
 ERROR_FILE=/tmp/uclient.error
 CERT="--ca-certificate=/etc/ssl/certs/download.crt"
+
+eval $(cat /etc/openwrt_release)
+
+# download file info
+FILE_INFO_JSON="$(/usr/lib/ddmesh/ddmesh-get-firmware-name.sh)"
+error=$(echo $FILE_INFO_JSON | jsonfilter -e '@.error')
+if [ -z "$FILE_INFO_JSON" -o -n "$error" ]; then
+	logger -s -t "$TAG" "Error: failed to download file information"
+	exit 1
+fi
 
 rm -f $ERROR_FILE
 
@@ -39,31 +41,19 @@ compare_versions() {
 }
 
 check_version() {
-	firmware_release_version=$(uclient-fetch $CERT -O - "$URL_RELEASE/version" 2>$ERROR_FILE)
-	[ -z "$firmware_release_version" ] && {
-		logger -t "$TAG" "Error: $(cat $ERROR_FILE)"
-		return 1
-	}
+	firmware_release_version=$(echo $FILE_INFO_JSON | jsonfilter -e '@.firmware_version')
 	firmware_current_version="$(cat /etc/version)"
 
-	logger -t "$TAG" "curver: $firmware_current_version, latest version: $firmware_release_version"
+	logger -s -t "$TAG" "curver: $firmware_current_version, latest version: $firmware_release_version"
 	compare_versions $firmware_release_version $firmware_current_version && return 0
 	return 1
 }
 
-get_server_md5sum() {
-	server_md5sum=$(uclient-fetch $CERT -O - "$URL_RELEASE/$URL_ARCH/md5sums" 2>$ERROR_FILE | grep "$FIRMWARE" | cut -d ' ' -f1)
-	[ -z "$server_md5sum" ] && {
-		logger -t "$TAG" "Error: $(cat $ERROR_FILE)"
-		return 1
-	}
-	return 0
-}
-
 download_firmware(){
-	logger -t "$TAG" "Try downloading $URL_RELEASE/$URL_ARCH/$FIRMWARE"
-	uclient-fetch $CERT -O $FIRMWARE_FILE "$URL_RELEASE/$URL_ARCH/$FIRMWARE" 2>$ERROR_FILE || {
-		logger -t "$TAG" "Error: $(cat $ERROR_FILE)"
+	url=$(echo $FILE_INFO_JSON | jsonfilter -e '@.url')
+	logger -s -t "$TAG" "Try downloading $url"
+	uclient-fetch $CERT -O $FIRMWARE_FILE "$url" 2>$ERROR_FILE || {
+		logger -s -t "$TAG" "Error: $(cat $ERROR_FILE)"
 		return 1
 	}
 	return 0
@@ -72,42 +62,53 @@ download_firmware(){
 
 case "$1" in
 	run)
-		logger -t "$TAG" "Auto Update Enabled: $enabled"
-		[ "$enabled" = "1" ] || exit 1
+		nightly=$2
+
+		autoupdate=$(echo $FILE_INFO_JSON | jsonfilter -e '@.fileinfo.autoupdate')
+
+		logger -s -t "$TAG" "Auto Update: Router enabled: $enabled; Server enabled: $autoupdate"
+		[ "$enabled" = "1" -a "$autoupdate" = "1" ] || exit 1
 
 		check_version || exit 1
-		logger -t "$TAG" "new version available"
+		logger -s -t "$TAG" "new version available"
 
 		download_firmware || exit 1
-		logger -t "$TAG" "firmware downloaded successfully"
+		logger -s -t "$TAG" "firmware downloaded successfully"
 
-		get_server_md5sum || exit 1
+		server_md5sum=$(echo $FILE_INFO_JSON | jsonfilter -e '@.fileinfo.md5sum')
 		file_md5sum=$(md5sum $FIRMWARE_FILE | cut -d' ' -f1)
 		if [ -z "$server_md5sum" -o "$server_md5sum" != "$file_md5sum" ]; then
-			logger -t "$TAG" "ERROR: Download md5sum failed !"
+			logger -s -t "$TAG" "ERROR: Download md5sum failed !"
 			rm -f $FIRMWARE_FILE
 			exit 1
 		fi
-		logger -t "$TAG" "md5sum is correct: $file_md5sum"
+		logger -s -t "$TAG" "md5sum is correct: $file_md5sum"
 
-		#check firmware (see /lib/upgrade)
-		if m=$(platform_check_image $FIRMWARE_FILE) ;then
+		#check firmware
+		if m=$(sysupgrade -T $FIRMWARE_FILE) ;then
 			#update configs after firmware update
 			uci set ddmesh.boot.boot_step=2
-			uci commit
+			# used for updated history
+			test -n "$nightly" && uci set ddmesh.boot.nightly_upgrade_running=1
+			# used to reset overlay md5sum
+			uci set ddmesh.boot.upgrade_running=1
+			uci_commit.sh
 			sync
-			logger -t "$TAG" "sysupgrade started..."
+			logger -s -t "$TAG" "sysupgrade started..."
 			sysupgrade $FIRMWARE_FILE 2>&1 >/dev/null &
 			exit 0
 		else # firmware check
 			rm -f $FIRMWARE_FILE
-			logger -t "$TAG" "ERROR: wrong firmware: $m"
+			logger -s -t "$TAG" "ERROR: wrong firmware: $m"
 			exit 1
 		fi
 		;;
 
 	check)
-		echo "auto update enabled: $enabled"
+		autoupdate=$(echo $FILE_INFO_JSON | jsonfilter -e '@.fileinfo.autoupdate')
+
+		echo "Auto Update: Router enabled: $enabled; Server enabled: $autoupdate"
+
 		check_version || echo "no new firmware"
 		echo "your version: $firmware_current_version"
 		echo "version on server: $firmware_release_version"
@@ -120,7 +121,7 @@ case "$1" in
 		;;
 
 	*)
-		echo "$0 <run | check | compare new old>"
+		echo "$0 <run [nightly] | check | compare new old>"
 		;;
 esac
